Add Spectacle functional tests
This commit adds basic functional tests for the [Spectacle plugin][1].
It will then be straightforward (in a subsequent pull request on the
Spectacle repo) to trigger these tests via a [GitHub Action][2] on every
push and pull request in the Spectacle repo.

This commit only adds a small number of functional tests with limited
coverage of Spectacle's functionality, but as these are the first
functional tests for Spectacle there is still significant
value from a regression prevention perspective.  Further tests can be
added in future commits if deemed useful.

The commit also provides a minimal framework for functional tests for
other plugins to be added in the future.  The Spectacle functional tests
in this commit serve as a useful example for anyone writing a future
plugin on how they can add functional tests.  To that end, it may
also be useful to make this gauge-tests repo a [template repository][3],
so that anyone writing a new Gauge plugin can easily use this repo as a
starting template for functional tests for their plugin.

[1]: https://github.com/getgauge/spectacle
[2]: https://github.com/features/actions
[3]: https://github.blog/2019-06-06-generate-new-repositories-with-repository-templates/

Signed-off-by: John Boyes <john@agilepathway.co.uk>